// src/lib/api/generated/models/namespace.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Hubuum REST API
 * OpenAPI documentation for the Hubuum REST service.
 * OpenAPI spec version: 0.1.0
 */

export interface Namespace {
  created_at: string;
  description: string;
  id: number;
  name: string;
  updated_at: string;
}